Move clean config option to the top level of generation config (#3165)
This PR targets branch `add-buf-gen-yaml-clean` and moves `clean` to
`buf.gen.yaml`'s top level, from the plugin level.

---------

+++ private/buf/bufgen/generator.go
@@ -34,6 +34,7 @@ import (
 	"github.com/bufbuild/buf/private/pkg/app"
 	"github.com/bufbuild/buf/private/pkg/command"
 	"github.com/bufbuild/buf/private/pkg/connectclient"
+	"github.com/bufbuild/buf/private/pkg/slicesext"
 	"github.com/bufbuild/buf/private/pkg/storage/storageos"
 	"github.com/bufbuild/buf/private/pkg/thread"
 	"github.com/bufbuild/buf/private/pkg/tracing"
@@ -104,13 +105,18 @@ func (g *generator) Generate(
 			return err
 		}
 	}
-	if err := g.deleteOuts(
-		ctx,
-		generateOptions.baseOutDirPath,
-		generateOptions.deleteOuts,
-		config.GeneratePluginConfigs(),
-	); err != nil {
-		return err
+	shouldDeleteOuts := config.CleanPluginOuts()
+	if generateOptions.deleteOuts != nil {
+		shouldDeleteOuts = *generateOptions.deleteOuts
+	}
+	if shouldDeleteOuts {
+		if err := g.deleteOuts(
+			ctx,
+			generateOptions.baseOutDirPath,
+			config.GeneratePluginConfigs(),
+		); err != nil {
+			return err
+		}
 	}
 	for _, image := range images {
 		if err := g.generateCode(
@@ -131,20 +137,21 @@ func (g *generator) Generate(
 func (g *generator) deleteOuts(
 	ctx context.Context,
 	baseOutDir string,
-	deleteAllOuts bool,
 	pluginConfigs []bufconfig.GeneratePluginConfig,
 ) error {
-	var pluginOuts []string
-	for _, pluginConfig := range pluginConfigs {
-		if deleteAllOuts || pluginConfig.Clean() {
-			if baseOutDir != "" && baseOutDir != "." {
-				pluginOuts = append(pluginOuts, filepath.Join(baseOutDir, pluginConfig.Out()))
-			} else {
-				pluginOuts = append(pluginOuts, pluginConfig.Out())
-			}
-		}
-	}
-	return bufprotopluginos.NewCleaner(g.storageosProvider).DeleteOuts(ctx, pluginOuts)
+	return bufprotopluginos.NewCleaner(g.storageosProvider).DeleteOuts(
+		ctx,
+		slicesext.Map(
+			pluginConfigs,
+			func(pluginConfig bufconfig.GeneratePluginConfig) string {
+				out := pluginConfig.Out()
+				if baseOutDir != "" && baseOutDir != "." {
+					return filepath.Join(baseOutDir, out)
+				}
+				return out
+			},
+		),
+	)
 }
 
 func (g *generator) generateCode(
@@ -481,7 +488,7 @@ func validateResponses(
 
 type generateOptions struct {
 	baseOutDirPath                string
-	deleteOuts                    bool
+	deleteOuts                    *bool
 	includeImportsOverride        *bool
 	includeWellKnownTypesOverride *bool
 }
